Add hot-patching instructions
